+++ guillotina_audit/__init__.py
@@ -12,7 +12,11 @@ app_settings = {
                 "index_permission_changes": False,
             },
         }
-    }
+    },
+    "commands": {
+        "audit-update-mappings": "guillotina_audit.commands.mappings.UpdateMappingsCommand",
+        "audit-update-settings": "guillotina_audit.commands.settings.UpdateSettingsCommand",
+    },
 }