+++ guillotina_audit/tests/test_audit_basic.py
@@ -14,13 +14,8 @@ pytestmark = pytest.mark.asyncio
 
 
 async def test_audit_basic(guillotina_es):
-    response, status = await guillotina_es(
-        "POST", "/db/guillotina/@addons", data=json.dumps({"id": "audit"})
-    )
-    assert status == 200
-    await asyncio.sleep(2)
-    audit_utility = query_utility(IAuditUtility)
     # Let's check the index has been created
+    audit_utility = query_utility(IAuditUtility)
     resp = await audit_utility.async_es.indices.get_alias()
     assert "audit" in resp
     resp = await audit_utility.async_es.indices.get_mapping(index="audit")
@@ -32,18 +27,23 @@ async def test_audit_basic(guillotina_es):
     )
     assert status == 201
     await asyncio.sleep(2)
-    resp, status = await guillotina_es("GET", "/db/guillotina/@audit")
+    resp, status = await guillotina_es(
+        "GET", "/db/guillotina/@audit?type_name=Container"
+    )
     assert status == 200
-    assert len(resp["hits"]["hits"]) == 2
+    assert len(resp["hits"]["hits"]) == 1
     assert resp["hits"]["hits"][0]["_source"]["action"] == "added"
     assert resp["hits"]["hits"][0]["_source"]["type_name"] == "Container"
     assert resp["hits"]["hits"][0]["_source"]["creator"] == "root"
     assert "title" in resp["hits"]["hits"][0]["_source"]["payload"]
 
-    assert resp["hits"]["hits"][1]["_source"]["action"] == "added"
-    assert resp["hits"]["hits"][1]["_source"]["type_name"] == "Item"
-    assert resp["hits"]["hits"][1]["_source"]["creator"] == "root"
-    assert "title" in resp["hits"]["hits"][1]["_source"]["payload"]
+    resp, status = await guillotina_es("GET", "/db/guillotina/@audit?type_name=Item")
+    assert status == 200
+
+    assert resp["hits"]["hits"][0]["_source"]["action"] == "added"
+    assert resp["hits"]["hits"][0]["_source"]["type_name"] == "Item"
+    assert resp["hits"]["hits"][0]["_source"]["creator"] == "root"
+    assert "title" in resp["hits"]["hits"][0]["_source"]["payload"]
 
     response, status = await guillotina_es("DELETE", "/db/guillotina/foo_item")
     await asyncio.sleep(2)
@@ -146,11 +146,6 @@ async def test_audit_basic(guillotina_es):
 
 
 async def test_audit_wildcard(guillotina_es):
-    response, status = await guillotina_es(
-        "POST", "/db/guillotina/@addons", data=json.dumps({"id": "audit"})
-    )
-    assert status == 200
-    await asyncio.sleep(2)
     audit_utility = query_utility(IAuditUtility)
 
     payload = AuditDocument(action="added", type_name="Fullscreen")
@@ -214,11 +209,6 @@ async def test_audit_wildcard(guillotina_es):
 
 
 async def test_json_dumps(guillotina_es):
-    response, status = await guillotina_es(
-        "POST", "/db/guillotina/@addons", data=json.dumps({"id": "audit"})
-    )
-    assert status == 200
-    await asyncio.sleep(2)
     audit_utility = query_utility(IAuditUtility)
     json.dumps(
         {"datetime": datetime.now(), "date": date.today()},
@@ -227,11 +217,6 @@ async def test_json_dumps(guillotina_es):
 
 
 async def test_permissions_modified_without_indexing(guillotina_es):
-    response, status = await guillotina_es(
-        "POST", "/db/guillotina/@addons", data=json.dumps({"id": "audit"})
-    )
-    assert status == 200
-
     response, status = await guillotina_es(
         "POST",
         "/db/guillotina/",
@@ -292,11 +277,6 @@ async def test_permissions_modified_without_indexing(guillotina_es):
 )
 async def test_permissions_modified_with_indexing(guillotina_es):
     response, status = await guillotina_es(
-        "POST", "/db/guillotina/@addons", data=json.dumps({"id": "audit"})
-    )
-    assert status == 200
-
-    response, status = await guillotina_es(
         "POST",
         "/db/guillotina/",
         data=json.dumps({"@type": "Item", "id": "foo_item", "title": "Foo Item"}),
@@ -329,3 +309,43 @@ async def test_permissions_modified_with_indexing(guillotina_es):
     # There should be one more document since indexing_permission_changes is True
     assert len(resp["hits"]["hits"]) == 3
     assert resp["hits"]["hits"][-1]["_source"]["action"] == "permissions_changed"
+
+
+async def test_metadata_field(guillotina_es):
+    audit_utility = query_utility(IAuditUtility)
+    payload = AuditDocument(
+        action="CreatingMetadata",
+        metadata={
+            "foo_number": 120,
+            "foo_string": "foo_string",
+            "foo_boolean": True,
+            "foo_list": [1, 2, 3, 4],
+            "foo_dict": {
+                "foo_number": 100,
+                "foo_string": "foo_string",
+                "foo_dict": {"foo_key": "foo_value"},
+            },
+            "foo_decimal": 1.234,
+        },
+    )
+    audit_utility.log_wildcard(payload)
+    # Let's check the index has been created
+    resp = await audit_utility.async_es.indices.get_alias()
+    await asyncio.sleep(2)
+    resp, status = await guillotina_es(
+        "GET",
+        "/db/guillotina/@audit?action=CreatingMetadata",
+    )
+    assert resp["hits"]["hits"][0]["_source"]["action"] == "CreatingMetadata"
+    assert resp["hits"]["hits"][0]["_source"]["metadata"] == {
+        "foo_boolean": True,
+        "foo_dict": {
+            "foo_dict": {"foo_key": "foo_value"},
+            "foo_number": 100,
+            "foo_string": "foo_string",
+        },
+        "foo_list": [1, 2, 3, 4],
+        "foo_number": 120,
+        "foo_string": "foo_string",
+        "foo_decimal": 1.234,
+    }